8-bit timers fixed

+++ modules/RAM/AVR/main.h
@@ -86,7 +86,7 @@ void set_claw(uint16_t width);
 /* Greatest 8-bit factor
  * Returns the greatest 8-bit factor of n
  */
-uint8_t G8BF(unsigned int n);
+uint8_t G8BF(uint16_t n);
 
 
 #endif
